requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
boto3==1.34.0
python-dotenv==1.0.0
werkzeug==3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
